Move Printl to console
